feat(refacto): remove useless redis and nginx changes + move changes into a different package
