+++ components/datadog/apps/nginx/ecs.go
@@ -2,56 +2,31 @@ package nginx
 
 import (
 	"github.com/DataDog/test-infra-definitions/common/config"
-	"github.com/DataDog/test-infra-definitions/common/namer"
 	"github.com/DataDog/test-infra-definitions/resources/aws"
-	"github.com/pulumi/pulumi-aws/sdk/v5/go/aws/alb"
+
 	classicECS "github.com/pulumi/pulumi-aws/sdk/v5/go/aws/ecs"
 	"github.com/pulumi/pulumi-awsx/sdk/go/awsx/awsx"
 	"github.com/pulumi/pulumi-awsx/sdk/go/awsx/ecs"
 	"github.com/pulumi/pulumi-awsx/sdk/go/awsx/lb"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
-	"os"
 )
 
 type EcsComponent struct {
 	pulumi.ResourceState
 }
 
-func GetEcsTargetGroupAndDns(e aws.Environment, namer namer.Namer, opts []pulumi.ResourceOption) (pulumi.StringOutput, pulumi.StringOutput, error) {
-	if os.Getenv("DD_DISABLE_LB") != "" {
-		lb_old, err := alb.NewLoadBalancer(e.Ctx, namer.ResourceName("lb"), &alb.LoadBalancerArgs{
-			LoadBalancerType: pulumi.StringPtr("application"),
-			Subnets:          pulumi.ToStringArray(e.DefaultSubnets()),
-			Internal:         pulumi.BoolPtr(true),
-			SecurityGroups:   pulumi.ToStringArray(e.DefaultSecurityGroups()),
-		}, opts...)
-		if err != nil {
-			return pulumi.StringOutput{}, pulumi.StringOutput{}, err
-		}
+func EcsAppDefinition(e aws.Environment, clusterArn pulumi.StringInput, opts ...pulumi.ResourceOption) (*EcsComponent, error) {
+	namer := e.Namer.WithPrefix("nginx")
+	opts = append(opts, e.WithProviders(config.ProviderAWS, config.ProviderAWSX))
 
-		tg, err := alb.NewTargetGroup(e.Ctx, namer.ResourceName("tg"), &alb.TargetGroupArgs{
-			Port:       pulumi.IntPtr(80),
-			Protocol:   pulumi.StringPtr("HTTP"),
-			TargetType: pulumi.StringPtr("instance"),
-			VpcId:      pulumi.StringPtr(e.DefaultVPCID()),
-		}, opts...)
-
-		if _, err = alb.NewListener(e.Ctx, namer.ResourceName("listener"), &alb.ListenerArgs{
-			LoadBalancerArn: lb_old.Arn,
-			Port:            pulumi.IntPtr(80),
-			Protocol:        pulumi.StringPtr("HTTP"),
-			DefaultActions: alb.ListenerDefaultActionArray{
-				&alb.ListenerDefaultActionArgs{
-					Type:           pulumi.String("forward"),
-					TargetGroupArn: tg.Arn,
-				},
-			},
-		}, opts...); err != nil {
-			return pulumi.StringOutput{}, pulumi.StringOutput{}, err
-		}
-		return tg.Arn, lb_old.DnsName, nil
+	ecsComponent := &EcsComponent{}
+	if err := e.Ctx.RegisterComponentResource("dd:apps", namer.ResourceName("grp"), ecsComponent, opts...); err != nil {
+		return nil, err
 	}
-	allb, err := lb.NewApplicationLoadBalancer(e.Ctx, namer.ResourceName("lb"), &lb.ApplicationLoadBalancerArgs{
+
+	opts = append(opts, pulumi.Parent(ecsComponent))
+
+	alb, err := lb.NewApplicationLoadBalancer(e.Ctx, namer.ResourceName("lb"), &lb.ApplicationLoadBalancerArgs{
 		Name:           e.CommonNamer.DisplayName(32, pulumi.String("nginx")),
 		SubnetIds:      e.RandomSubnets(),
 		Internal:       pulumi.BoolPtr(true),
@@ -68,27 +43,6 @@ func GetEcsTargetGroupAndDns(e aws.Environment, namer namer.Namer, opts []pulumi
 			Protocol: pulumi.StringPtr("HTTP"),
 		},
 	}, opts...)
-	if err != nil {
-		return pulumi.StringOutput{}, pulumi.StringOutput{}, err
-	}
-	return allb.DefaultTargetGroup.Arn(), allb.LoadBalancer.DnsName(), nil
-
-}
-
-func EcsAppDefinition(e aws.Environment, clusterArn pulumi.StringInput, opts ...pulumi.ResourceOption) (*EcsComponent, error) {
-	namer := e.Namer.WithPrefix("nginx")
-	opts = append(opts, e.WithProviders(config.ProviderAWS, config.ProviderAWSX))
-
-	ecsComponent := &EcsComponent{}
-	if err := e.Ctx.RegisterComponentResource("dd:apps", namer.ResourceName("grp"), ecsComponent, opts...); err != nil {
-		return nil, err
-	}
-
-	opts = append(opts, pulumi.Parent(ecsComponent))
-	var targetGroup pulumi.StringOutput
-	var Dnsname pulumi.StringOutput
-	var err error
-	targetGroup, Dnsname, err = GetEcsTargetGroupAndDns(e, namer, opts)
 	if err != nil {
 		return nil, err
 	}
@@ -171,7 +125,7 @@ func EcsAppDefinition(e aws.Environment, clusterArn pulumi.StringInput, opts ...
 			&classicECS.ServiceLoadBalancerArgs{
 				ContainerName:  pulumi.String("nginx"),
 				ContainerPort:  pulumi.Int(80),
-				TargetGroupArn: targetGroup,
+				TargetGroupArn: alb.DefaultTargetGroup.Arn(),
 			},
 		},
 	}, opts...); err != nil {
@@ -190,7 +144,7 @@ func EcsAppDefinition(e aws.Environment, clusterArn pulumi.StringInput, opts ...
 					Image: pulumi.String("ghcr.io/datadog/apps-http-client:main"),
 					Command: pulumi.StringArray{
 						pulumi.String("-url"),
-						pulumi.Sprintf("http://%s", Dnsname),
+						pulumi.Sprintf("http://%s", alb.LoadBalancer.DnsName()),
 					},
 					Cpu:    pulumi.IntPtr(50),
 					Memory: pulumi.IntPtr(32),
