+++ go.mod
@@ -1,6 +1,8 @@
 module github.com/DataDog/test-infra-definitions
 
-go 1.20
+go 1.21
+
+toolchain go1.21.1
 
 require (
 	github.com/Masterminds/semver v1.5.0
